requirements.txt
Flask==3.0.2
requests==2.31.0
pydantic==2.6.4
gunicorn==21.2.0
python-dotenv==1.0.1
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1